完善頁面功能(新增支出、刪除支出、修改支出)，版面配置更改

+++ input.py
@@ -16,7 +16,7 @@ def add_expense():
     # 2. 輸入金額
     while True:
         try:
-            amount = float(input("請輸入金額: "))
+            amount = int(input("請輸入金額: "))
             break
         except ValueError:
             print("金額格式錯誤，請輸入數字。")
